refactor(zfs): extract ZFSPoolsTable inline components and row builders
Sub-table components (PoolSubTable, VdevDiskSubTable, DiskSubTable) are now
in components/zfs/subtables/. Row builder functions are in
components/zfs/utils/zfs-row-builders.ts.
