+++ src/components/zfs/ZFSPoolsTable.tsx
@@ -1,15 +1,17 @@
-import { memo, useCallback, useMemo, type ReactNode } from 'react';
+import { useCallback, useMemo } from 'react';
 import type { ColumnDef, ExpandedState } from '@tanstack/react-table';
 import { Box, CircularProgress, Typography } from '@mui/material';
 import { StaleDataAlert } from '@/components/shared-table/StaleDataAlert';
 import { DataTable, type MetricGroup } from '@/components/shared-table/DataTable';
 import { metricColumn, nameColumn } from '@/components/shared-table/columns';
 import { EMPTY_METRIC } from '@/components/shared-table';
-import type { ZFSStatsRow, ZFSHostHierarchy, ZFSHostStats, PoolStats, VdevStats } from '@/types/zfs';
+import type { ZFSStatsRow, ZFSHostHierarchy } from '@/types/zfs';
 import { buildZFSHostHierarchy } from '@/lib/utils/zfs-hierarchy-builder';
 import { formatBytesParts, formatAsPercentParts } from '@/formatters/metrics';
 import { useGeneralSettings, useZfsSettings } from '@/hooks/useSettings';
 import ZFSEntityCell from '@/components/zfs/ZFSEntityCell';
+import PoolSubTable from '@/components/zfs/subtables/PoolSubTable';
+import { buildHostRow } from '@/components/zfs/utils/zfs-row-builders';
 
 /** Flattened row model for the DataTable tree structure */
 export interface ZFSTableRow {
@@ -271,323 +273,4 @@ export default function ZFSPoolsTable({
       />
     </Box>
   );
-}
-
-/**
- * Nested DataTable for pool rows within an expanded host.
- * Each pool with expandable children renders a VdevDiskSubTable via renderDetailPanel.
- */
-const PoolSubTable = memo(function PoolSubTable({
-  pools,
-  columns,
-  isPoolExpanded,
-  togglePoolExpanded,
-  isVdevExpanded,
-  toggleVdevExpanded,
-}: Readonly<{
-  pools: ZFSTableRow[];
-  columns: ColumnDef<ZFSTableRow, unknown>[];
-  isPoolExpanded: (id: string, totalPools: number) => boolean;
-  togglePoolExpanded: (id: string) => void;
-  isVdevExpanded: (id: string) => boolean;
-  toggleVdevExpanded: (id: string) => void;
-}>) {
-  const poolExpanded = useMemo<ExpandedState>(() => {
-    const state: Record<string, boolean> = {};
-    for (const pool of pools) {
-      if (pool.canExpand && pool.totalPools != null) {
-        state[pool.id] = isPoolExpanded(pool.id, pool.totalPools);
-      }
-    }
-    return state;
-  }, [pools, isPoolExpanded]);
-
-  const handlePoolExpandedChange = useCallback(
-    (newExpanded: ExpandedState) => {
-      if (typeof newExpanded === 'boolean') return;
-      const current = poolExpanded as Record<string, boolean>;
-      for (const [id, value] of Object.entries(newExpanded)) {
-        if (value !== (current[id] ?? false)) {
-          togglePoolExpanded(id);
-        }
-      }
-      for (const id of Object.keys(current)) {
-        if (current[id] && !(id in newExpanded)) {
-          togglePoolExpanded(id);
-        }
-      }
-    },
-    [poolExpanded, togglePoolExpanded],
-  );
-
-  /** Pool detail panel: a nested DataTable of vdev/disk rows */
-  const renderPoolDetail = useCallback(
-    (row: ZFSTableRow): ReactNode => {
-      if (!row.canExpand || !row.children?.length) return null;
-      return (
-        <VdevDiskSubTable
-          vdevsAndDisks={row.children}
-          columns={columns}
-          isVdevExpanded={isVdevExpanded}
-          toggleVdevExpanded={toggleVdevExpanded}
-        />
-      );
-    },
-    [columns, isVdevExpanded, toggleVdevExpanded],
-  );
-
-  return (
-    <DataTable
-      data={pools}
-      columns={columns}
-      getRowId={(row) => row.id}
-      renderDetailPanel={renderPoolDetail}
-      expandedState={poolExpanded}
-      onExpandedChange={handlePoolExpandedChange}
-      enableSorting={false}
-      showHeader={false}
-    />
-  );
-});
-
-/**
- * Nested DataTable for vdev and disk rows within an expanded pool.
- * Vdevs with disk children render a DiskSubTable via renderDetailPanel.
- */
-const VdevDiskSubTable = memo(function VdevDiskSubTable({
-  vdevsAndDisks,
-  columns,
-  isVdevExpanded,
-  toggleVdevExpanded,
-}: Readonly<{
-  vdevsAndDisks: ZFSTableRow[];
-  columns: ColumnDef<ZFSTableRow, unknown>[];
-  isVdevExpanded: (id: string) => boolean;
-  toggleVdevExpanded: (id: string) => void;
-}>) {
-  const vdevExpanded = useMemo<ExpandedState>(() => {
-    const state: Record<string, boolean> = {};
-    for (const row of vdevsAndDisks) {
-      if (row.type === 'vdev' && row.canExpand) {
-        state[row.id] = isVdevExpanded(row.id);
-      }
-    }
-    return state;
-  }, [vdevsAndDisks, isVdevExpanded]);
-
-  const handleVdevExpandedChange = useCallback(
-    (newExpanded: ExpandedState) => {
-      if (typeof newExpanded === 'boolean') return;
-      const current = vdevExpanded as Record<string, boolean>;
-      for (const [id, value] of Object.entries(newExpanded)) {
-        if (value !== (current[id] ?? false)) {
-          toggleVdevExpanded(id);
-        }
-      }
-      for (const id of Object.keys(current)) {
-        if (current[id] && !(id in newExpanded)) {
-          toggleVdevExpanded(id);
-        }
-      }
-    },
-    [vdevExpanded, toggleVdevExpanded],
-  );
-
-  /** Vdev detail panel: a nested DataTable of disk rows */
-  const renderVdevDetail = useCallback(
-    (row: ZFSTableRow): ReactNode => {
-      if (row.type !== 'vdev' || !row.canExpand || !row.children?.length) return null;
-      return (
-        <DiskSubTable
-          disks={row.children}
-          columns={columns}
-        />
-      );
-    },
-    [columns],
-  );
-
-  return (
-    <DataTable
-      data={vdevsAndDisks}
-      columns={columns}
-      getRowId={(row) => row.id}
-      renderDetailPanel={renderVdevDetail}
-      expandedState={vdevExpanded}
-      onExpandedChange={handleVdevExpandedChange}
-      enableSorting={false}
-      showHeader={false}
-    />
-  );
-});
-
-/**
- * Leaf-level DataTable for disk rows within an expanded vdev.
- * No expansion, just renders disk rows.
- */
-const DiskSubTable = memo(function DiskSubTable({
-  disks,
-  columns,
-}: Readonly<{
-  disks: ZFSTableRow[];
-  columns: ColumnDef<ZFSTableRow, unknown>[];
-}>) {
-  return (
-    <DataTable
-      data={disks}
-      columns={columns}
-      getRowId={(row) => row.id}
-      enableSorting={false}
-      showHeader={false}
-    />
-  );
-});
-
-/**
- * Build a host-level tree row with pool children from the ZFS hierarchy.
- */
-function buildHostRow(host: ZFSHostStats, totalHosts: number): ZFSTableRow {
-  const a = host.aggregated;
-  const totalPools = host.pools.size;
-  const sortedPools = Array.from(host.pools.values()).sort((a, b) =>
-    a.data.name.localeCompare(b.data.name),
-  );
-
-  const children = sortedPools.map((pool) => buildPoolRow(pool, totalPools));
-
-  return {
-    type: 'host',
-    id: `host:${host.hostName}`,
-    name: host.hostName,
-    indent: 0,
-    capacityAlloc: a.capacityAlloc,
-    capacityFree: a.capacityFree,
-    readOpsPerSec: a.readOpsPerSec,
-    writeOpsPerSec: a.writeOpsPerSec,
-    readBytesPerSec: a.readBytesPerSec,
-    writeBytesPerSec: a.writeBytesPerSec,
-    badge: { label: `${a.poolCount} pool${a.poolCount === 1 ? '' : 's'}` },
-    canExpand: totalHosts > 1 && totalPools > 0,
-    totalHosts,
-    children,
-  };
-}
-
-/**
- * Build a pool-level tree row with vdev/disk children.
- * Preserves the original table's badge and vdev-skip logic.
- */
-function buildPoolRow(pool: PoolStats, totalPools: number): ZFSTableRow {
-  const vdevs = Array.from(pool.vdevs.values()).sort((a, b) =>
-    a.data.name.localeCompare(b.data.name),
-  );
-  const individualDisks = Array.from(pool.individualDisks.values()).sort((a, b) =>
-    a.data.name.localeCompare(b.data.name),
-  );
-
-  const singleVdev = vdevs.length === 1 && individualDisks.length === 0;
-  const isSingleDiskPool =
-    (singleVdev && vdevs[0].disks.size <= 1) ||
-    (vdevs.length === 0 && individualDisks.length === 1);
-  const isSingleVdevMultiDisk = singleVdev && vdevs[0].disks.size > 1;
-
-  let badge: { label: string; tooltip?: string } | undefined;
-  if (isSingleDiskPool) {
-    const tooltipName = singleVdev
-      ? Array.from(vdevs[0].disks.values())[0]?.data.name ?? vdevs[0].data.name
-      : individualDisks[0]?.data.name;
-    badge = { label: 'single disk', tooltip: tooltipName };
-  } else if (singleVdev) {
-    badge = { label: vdevs[0].data.name };
-  }
-
-  const expandable = !isSingleDiskPool;
-  let children: ZFSTableRow[] | undefined;
-
-  if (expandable) {
-    if (isSingleVdevMultiDisk) {
-      // Skip the vdev level, show disks directly under pool
-      const vdevDisks = Array.from(vdevs[0].disks.values()).sort((a, b) =>
-        a.data.name.localeCompare(b.data.name),
-      );
-      children = vdevDisks.map((disk) => buildDiskRow(disk.data.id, disk.data.name, disk.data, 0));
-    } else {
-      const vdevChildren = vdevs.map((vdev) => buildVdevRow(vdev));
-      const diskChildren = individualDisks.map((disk) =>
-        buildDiskRow(disk.data.id, disk.data.name, disk.data, 0),
-      );
-      children = [...vdevChildren, ...diskChildren];
-    }
-  }
-
-  return {
-    type: 'pool',
-    id: pool.data.id,
-    name: pool.data.name,
-    indent: 0,
-    capacityAlloc: pool.data.capacity.alloc,
-    capacityFree: pool.data.capacity.free,
-    readOpsPerSec: pool.data.rates.readOpsPerSec,
-    writeOpsPerSec: pool.data.rates.writeOpsPerSec,
-    readBytesPerSec: pool.data.rates.readBytesPerSec,
-    writeBytesPerSec: pool.data.rates.writeBytesPerSec,
-    utilizationPercent: pool.data.rates.utilizationPercent,
-    badge,
-    canExpand: expandable,
-    totalPools,
-    children,
-  };
-}
-
-/**
- * Build a vdev-level tree row with disk children.
- */
-function buildVdevRow(vdev: VdevStats): ZFSTableRow {
-  const sortedDisks = Array.from(vdev.disks.values()).sort((a, b) =>
-    a.data.name.localeCompare(b.data.name),
-  );
-
-  const hasDisks = sortedDisks.length > 0;
-  const children = hasDisks
-    ? sortedDisks.map((disk) => buildDiskRow(disk.data.id, disk.data.name, disk.data, 0))
-    : undefined;
-
-  return {
-    type: 'vdev',
-    id: `vdev:${vdev.data.id}`,
-    name: vdev.data.name,
-    indent: 0,
-    capacityAlloc: vdev.data.capacity.alloc > 0 ? vdev.data.capacity.alloc : undefined,
-    capacityFree: vdev.data.capacity.alloc > 0 ? vdev.data.capacity.free : undefined,
-    readOpsPerSec: vdev.data.rates.readOpsPerSec,
-    writeOpsPerSec: vdev.data.rates.writeOpsPerSec,
-    readBytesPerSec: vdev.data.rates.readBytesPerSec,
-    writeBytesPerSec: vdev.data.rates.writeBytesPerSec,
-    utilizationPercent: vdev.data.rates.utilizationPercent,
-    canExpand: hasDisks,
-    children,
-  };
-}
-
-/**
- * Build a disk leaf row (no children).
- */
-function buildDiskRow(
-  id: string,
-  name: string,
-  data: { rates: { readOpsPerSec: number; writeOpsPerSec: number; readBytesPerSec: number; writeBytesPerSec: number; utilizationPercent: number } },
-  indent: number,
-): ZFSTableRow {
-  return {
-    type: 'disk',
-    id: `disk:${id}`,
-    name,
-    indent,
-    readOpsPerSec: data.rates.readOpsPerSec,
-    writeOpsPerSec: data.rates.writeOpsPerSec,
-    readBytesPerSec: data.rates.readBytesPerSec,
-    writeBytesPerSec: data.rates.writeBytesPerSec,
-    utilizationPercent: data.rates.utilizationPercent,
-    canExpand: false,
-  };
 }
